Changed a few things: Updated Indexes, URL to filepath instead, Checks all words read instead of only one .
Updated Indexes, 
URL to filepath instead, 
Checks all words read instead of only one .

Things to Change:
- As Zach mentioned: we should create equivalences or similar words.

I noticed that sometimes the model registered an extra letter, making it incorrect. Should we let this pass?

The data set has some images that are intentionally in the wrong place that was meant for training a CNN. Ex: Arkansas license plate in Alabama folder. 
IDK what to do here.

For Guam:
GUAM USA is read as a single entity. So it is processed as incorrect.

Probably some more missed. Should we ignore?