Added equivalencies and regex
Now iterates through all equivalencies for each plate until it finds one that is contained within the words extracted from the plate.  
Will check to see if the plate it thought it was based on the equivalency is actually from the right state by comparing the indices from the csv and the list of equivalencies.  Number correct will only be incremented if it was actually correct, as just finding an equivalency does not necessarily mean that the state was correct (i.e. Washington D.C. will probably get identified as Washington most times, same with West Virginia and Virginia)